% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{myttest function will look for a difference in the population variances and see whether or not the data.
If the data supports the conculusion that the variances are different or if the variances are paired
then the t-test will be ran and will return a list.}
\usage{
myttest(x, y, alpha, paired)
}
\arguments{
\item{x}{data}

\item{y}{data}

\item{alpha}{ci}

\item{paired}{paired parameter}
}
\value{
A data frame
}
\description{
- The one following type of test could be ran: "Welch", " T-test", "Paired"
- A decision on whether or not to reject the null of equality of the mean at the alpha level of significance.
- A summary of the chosen t test.
}
