% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{This function will return a a boxplot of the data from the Rttest class, it will return multiple boxplots
One when the non-paired data is analyzed
Second one with the differences in data when the paired is True.
Third one is the ci for the difference of means?}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{data from myttest function}

\item{...}{Extra parameters}
}
\value{
Multiple plots of the categories and samples
}
\description{
This function will return a a boxplot of the data from the Rttest class, it will return multiple boxplots
One when the non-paired data is analyzed
Second one with the differences in data when the paired is True.
Third one is the ci for the difference of means?
}
