% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{This method prints to the command line the confidence interval for muX and muY.}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{list from myttest function}

\item{...}{extra parameters}
}
\value{
This function will return a list of summary data with the type of the t-test,
confidence interval of Rttest.
}
\description{
This method prints to the command line the confidence interval for muX and muY.
}
